backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi[standard]>=0.115.11",
    "pyyaml>=6.0.2",
    "sqlmodel>=0.0.24",
]
